refactor: вынести обработчики команд и логику задач в отдельные файлы
- Вынес обработчики команд Telegram из функции Run в отдельные функции
- Создал internal/bot/task.go для работы с задачами (Task)
- Улучшил читаемость и поддерживаемость кода
- Все тесты проходят
