+++ go-project/README.md
@@ -1,0 +1,43 @@
+# Go Project
+
+This is a simple Go project that demonstrates the structure and organization of a Go application.
+
+## Project Structure
+
+```
+go-project
+├── cmd
+│   └── main.go          # Entry point of the application
+├── internal
+│   └── app
+│       └── app.go      # Application lifecycle management
+├── pkg
+│   └── utils.go        # Utility functions
+├── go.mod              # Module definition
+└── go.sum              # Dependency checksums
+```
+
+## Getting Started
+
+To get started with this project, clone the repository and navigate into the project directory:
+
+```bash
+git clone <repository-url>
+cd go-project
+```
+
+## Running the Application
+
+To run the application, use the following command:
+
+```bash
+go run cmd/main.go
+```
+
+## Contributing
+
+Contributions are welcome! Please feel free to submit a pull request or open an issue for any suggestions or improvements.
+
+## License
+
+This project is licensed under the MIT License. See the LICENSE file for more details.
