+++ go-project/go.mod
@@ -1,0 +1,7 @@
+module go-project
+
+go 1.18
+
+require ( 
+    // Add any dependencies here
+)
