+++ go-project/internal/app/app.go
@@ -1,0 +1,13 @@
+package app
+
+type App struct{}
+
+func (a *App) Start() error {
+    // Logic to start the application
+    return nil
+}
+
+func (a *App) Stop() error {
+    // Logic to stop the application
+    return nil
+}
